move meta commands to their own package
