--- datadog_checks_dev/datadog_checks/dev/tooling/commands/meta_commands/__init__.py
@@ -1,8 +1,0 @@
-# (C) Datadog, Inc. 2018
-# All rights reserved
-# Licensed under a 3-clause BSD style license (see LICENSE)
-from .prometheus import prom
-
-ALL_COMMANDS = (
-    prom,
-)
